backend/requirements.txt
fastapi==0.115.7
uvicorn[standard]==0.35.0
SQLAlchemy==2.0.37
psycopg2-binary==2.9.10
python-multipart==0.0.20
python-docx==1.1.2
minio==7.2.15
pydantic-settings==2.7.1
